rearranged and organized folder structure for clientlayout
