+++ client/src/components/clientlayout/login/Login.css
@@ -1,0 +1,7 @@
+.full-spread {
+  background-image: radial-gradient(white, gray);
+  height: 100vh;
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+}
