+++ client/src/components/clientlayout/mainpage/truckmain.jpg
