+++ client/src/components/clientlayout/navbar/Navbar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import logo from "./images/logo.png";
+import logo from "./logo.png";
 
 function Navbar() {
   return (
