+++ client/src/components/clientlayout/navbar/logo.png
