+++ client/src/components/clientlayout/signup/Signup.js
@@ -1,11 +1,11 @@
 import React from "react";
-import "./Main.css";
+import "./Signup.css";
 
 function MainPage() {
   return (
     <div>
       <div className="full-spread d-flex justify-content-center align-items-center">
-        <h1>Log in</h1>
+        <h1>Sign up</h1>
       </div>
     </div>
   );
