+++ client/src/components/clientlayout/zone/Spot1.js
