+++ client/src/components/clientlayout/zone/Spot2.js
