--- client/src/components/layouts/Main.css
@@ -1,19 +1,0 @@
-.hero-image {
-  background-image: url(../layouts/images/truckmain.jpg);
-  height: 100vh;
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center;
-}
-
-.full-spread {
-  background-image: radial-gradient(white, gray);
-  height: 100vh;
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center;
-}
-
-.maintext {
-  background-color: rgba(rgb(131, 131, 131), 0.5) !important;
-}
