--- client/src/components/layouts/navbar.css
@@ -1,4 +1,0 @@
-ul li {
-  text-decoration: none;
-  text-decoration-style: none;
-}
